// src/app/Models/logistic-zone.ts
export class LogisticZone {
    createdBy?: string;
    createdDate?: string;
    descr?: string;
    description?: string;
    effectiveDate?: string;
    expiryDate?: string;
    effectiveDt?:string;
    expDt?: string;
    id?: number;
    status?: number;
    updateDate?: string;
    updatedBy?: string;
    zoneName?: string;

    // option
    states:any;
}